trunk/bsp_device_driver/pi/requirements.txt
cbor2==5.7.1
luma.core==2.5.3
luma.oled==3.14.0
pillow==12.0.0
RPi.GPIO==0.7.1
smbus2==0.6.0
spidev==3.8
